refactor: remove UserRole entity
RoleType covers our needs for user.role no need to keep it in database. It would be meaningful if we were getting permissions based on roles which are created by system admins after the app is deployed. We do not have such a requirement here.

I did this also to give an example of how migrations help us to make such complex refactorings in our code and db schema.